remove references to findfiles2New and aiTextSearchProviderNew
