+++ src/vscode-dts/vscode.proposed.aiTextSearchProvider.d.ts
@@ -4,38 +4,6 @@
  *--------------------------------------------------------------------------------------------*/
 
 declare module 'vscode' {
-
-	/**
-	 * Options that apply to AI text search.
-	 */
-	export interface AITextSearchOptions extends SearchOptions {
-		/**
-		 * The maximum number of results to be returned.
-		 */
-		maxResults: number;
-
-		/**
-		 * Options to specify the size of the result text preview.
-		 */
-		previewOptions?: TextSearchPreviewOptions;
-
-		/**
-		 * Exclude files larger than `maxFileSize` in bytes.
-		 */
-		maxFileSize?: number;
-
-		/**
-		 * Number of lines of context to include before each match.
-		 */
-		beforeContext?: number;
-
-		/**
-		 * Number of lines of context to include after each match.
-		 */
-		afterContext?: number;
-	}
-
-
 	/**
 	 * An AITextSearchProvider provides additional AI text search results in the workspace.
 	 */
@@ -46,14 +14,14 @@ declare module 'vscode' {
 		readonly name?: string;
 
 		/**
+		 *
 		 * Provide results that match the given text pattern.
 		 * @param query The parameter for this query.
 		 * @param options A set of options to consider while searching.
 		 * @param progress A progress callback that must be invoked for all results.
 		 * @param token A cancellation token.
 		 */
-		provideAITextSearchResults(query: string, options: AITextSearchOptions, progress: Progress<TextSearchResult>, token: CancellationToken): ProviderResult<TextSearchComplete>;
-
+		provideAITextSearchResults(query: string, options: TextSearchProviderOptions, progress: Progress<TextSearchResultNew>, token: CancellationToken): ProviderResult<TextSearchCompleteNew>;
 	}
 
 	export namespace workspace {
