+++ src/vscode-dts/vscode.proposed.findFiles2.d.ts
@@ -6,84 +6,100 @@
 declare module 'vscode' {
 
 	export interface FindFiles2Options {
-		// note: this is just FindTextInFilesOptions without select properties (include, previewOptions, beforeContext, afterContext)
-
 		/**
-		 * A {@link GlobPattern glob pattern} that defines files and folders to exclude. The glob pattern
-		 * will be matched against the file paths of resulting matches relative to their workspace.
+		 * An array of {@link GlobPattern GlobPattern} that defines files to exclude.
+		 * The glob patterns will be matched against the file paths of files relative to their workspace or {@link RelativePattern.baseUri} if applicable.
+		 *
+		 * If more than one value is used, the values are combined with a logical AND.
+		 * For example, consider the following code:
+		 *
+		 * ```ts
+		 * const ab = findFiles(['**​/*.js'], {exclude: ['*.ts', '*.js']});
+		 * const a = findFiles(['**​/*.js'], {exclude: ['*.ts']});
+		 * const b = findFiles(['**​/*.js'], {exclude: ['*.js']});
+		 * ```
+		 *
+		 * In this, `ab` will be the intersection of results from `a` and `b`.
 		 */
-		exclude?: GlobPattern;
+		exclude?: GlobPattern[];
 
 		/**
-		 * Whether to use the values for files.exclude. Defaults to true.
+		 * Which settings to follow when searching for files. Defaults to {@link ExcludeSettingOptions.searchAndFilesExclude}.
 		 */
-		useDefaultExcludes?: boolean;
+		useExcludeSettings?: ExcludeSettingOptions;
 
 		/**
-		 * Whether to use the values for search.exclude. Defaults to true. Will not be followed if `useDefaultExcludes` is set to `false`.
-		 */
-		useDefaultSearchExcludes?: boolean;
-
-		/**
-		 * The maximum number of results to search for
+		 * The maximum number of results to search for. Defaults to 20000 results.
 		 */
 		maxResults?: number;
 
 		/**
-		 * Whether external files that exclude files, like .gitignore, should be respected.
-		 * Defaults to the value for `search.useIgnoreFiles` in settings.
-		 * For more info, see the setting listed above.
+		 * Which file locations have ignore (`.gitignore` or `.ignore`) files to follow.
+		 *
+		 * When any of these fields are `undefined`, the value will either be assumed (e.g. if only one is valid),
+		 * or it will follow settings based on the corresponding `search.use*IgnoreFiles` setting.
+		 *
+		 * Will log an error if an invalid combination is set.
+		 *
+		 * Although `.ignore` files are uncommon, they can be leveraged if there are patterns
+		 * that should not be known to git, but should be known to the search providers.
+		 * They should be in the same locations where `.gitignore` files are found, and they follow the same format.
 		 */
-		useIgnoreFiles?: boolean;
-
-		/**
-		 * Whether global files that exclude files, like .gitignore, should be respected.
-		 * Must set `useIgnoreFiles` to `true` to use this.
-		 * Defaults to the value for `search.useGlobalIgnoreFiles` in settings.
-		 * For more info, see the setting listed above.
-		 */
-		useGlobalIgnoreFiles?: boolean;
-
-		/**
-		 * Whether files in parent directories that exclude files, like .gitignore, should be respected.
-		 * Must set `useIgnoreFiles` to `true` to use this.
-		 * Defaults to the value for `search.useParentIgnoreFiles` in settings.
-		 * For more info, see the setting listed above.
-		 */
-		useParentIgnoreFiles?: boolean;
+		useIgnoreFiles?: {
+			/**
+			 * Use ignore files at the current workspace root.
+			 * May default to `search.useIgnoreFiles` setting if not set.
+			 */
+			local?: boolean;
+			/**
+			 * Use ignore files at the parent directory. When set to `true`, {@link FindFiles2OptionsNew.useIgnoreFiles.local} must also be `true`.
+			 * May default to `search.useParentIgnoreFiles` setting if not set.
+			 */
+			parent?: boolean;
+			/**
+			 * Use global ignore files. When set to `true`, {@link FindFiles2OptionsNew.useIgnoreFiles.local} must also be `true`.
+			 * May default to `search.useGlobalIgnoreFiles` setting if not set.
+			 */
+			global?: boolean;
+		};
 
 		/**
 		 * Whether symlinks should be followed while searching.
 		 * Defaults to the value for `search.followSymlinks` in settings.
-		 * For more info, see the setting listed above.
+		 * For more info, see the setting description for `search.followSymlinks`.
 		 */
 		followSymlinks?: boolean;
-
-		/**
-		 * If set to true, the `filePattern` arg will be fuzzy-searched instead of glob-searched.
-		 * If `filePattern` is a {@link RelativePattern relative pattern}, then the fuzzy search will act on the `pattern` of the {@link RelativePattern RelativePattern}
-		 */
-		fuzzy?: boolean;
 	}
 
-	/**
-	 * Represents a session of a currently logged in user.
-	 */
 	export namespace workspace {
 		/**
+		 * WARNING: VERY EXPERIMENTAL.
+		 *
 		 * Find files across all {@link workspace.workspaceFolders workspace folders} in the workspace.
 		 *
 		 * @example
-		 * findFiles('**​/*.js', {exclude: '**​/out/**', useIgnoreFiles: true, maxResults: 10})
+		 * findFiles(['**​/*.js'], {exclude: ['**​/out/**'], useIgnoreFiles: true, maxResults: 10})
 		 *
-		 * @param filePattern A {@link GlobPattern glob pattern} that defines the files to search for. The glob pattern
-		 * will be matched against the file paths of resulting matches relative to their workspace. Use a {@link RelativePattern relative pattern}
-		 * to restrict the search results to a {@link WorkspaceFolder workspace folder}.
+		 * @param filePattern An array of {@link GlobPattern GlobPattern} that defines the files to search for.
+		 * The glob patterns will be matched against the file paths of files relative to their workspace or {@link baseUri GlobPattern.baseUri} if applicable.
+		 * Use a {@link RelativePattern RelativePatten} to restrict the search results to a {@link WorkspaceFolder workspace folder}.
+		 *
+		 * If more than one value is used, the values are combined with a logical OR.
+		 *
+		 * For example, consider the following code:
+		 *
+		 * ```ts
+		 * const ab = findFiles(['*.ts', '*.js']);
+		 * const a = findFiles(['**​/*.ts']);
+		 * const b = findFiles(['**​/*.js']);
+		 * ```
+		 *
+		 * In this, `ab` will be the union of results from `a` and `b`.
 		 * @param options A set of {@link FindFiles2Options FindFiles2Options} that defines where and how to search (e.g. exclude settings).
 		 * @param token A token that can be used to signal cancellation to the underlying search engine.
 		 * @returns A thenable that resolves to an array of resource identifiers. Will return no results if no
 		 * {@link workspace.workspaceFolders workspace folders} are opened.
 		 */
-		export function findFiles2(filePattern: GlobPattern, options?: FindFiles2Options, token?: CancellationToken): Thenable<Uri[]>;
+		export function findFiles2(filePattern: GlobPattern[], options?: FindFiles2Options, token?: CancellationToken): Thenable<Uri[]>;
 	}
 }
